refactor(dashboard): decompose DashboardService into Overview/Collections/Ops compute services
Behavior-preserving, pure-read. Facade keeps the cache layer (cached() helper +
exact keys/TTLs) + 12 thin public wrappers; heavy compute moved VERBATIM into 3
domain sub-services (DashboardOverviewService/Collections/Ops) injecting prisma.
Facade sheds prisma. 1074->~150 LOC facade. Controller + external consumers
unaffected (surface preserved). tests green, tsc 0, lint 0.
