+++ apps/api/src/modules/dashboard/dashboard.service.ts
@@ -1,17 +1,19 @@
 import { Injectable, Inject } from '@nestjs/common';
 import { CACHE_MANAGER } from '@nestjs/cache-manager';
 import { Cache } from 'cache-manager';
-import { PrismaService } from '../../prisma/prisma.service';
-import { Prisma } from '@prisma/client';
-import { calculateDaysOverdue, calculateDaysElapsed } from '../../utils/date.util';
 import { StructuredLoggerService } from '../../common/logger';
+import { DashboardOverviewService } from './services/dashboard-overview.service';
+import { DashboardCollectionsService } from './services/dashboard-collections.service';
+import { DashboardOpsService } from './services/dashboard-ops.service';
 
 @Injectable()
 export class DashboardService {
   private readonly structuredLogger = new StructuredLoggerService(DashboardService.name);
   constructor(
-    private prisma: PrismaService,
     @Inject(CACHE_MANAGER) private cache: Cache,
+    private overview: DashboardOverviewService,
+    private collections: DashboardCollectionsService,
+    private ops: DashboardOpsService,
   ) {}
 
   /**
@@ -54,467 +56,56 @@ export class DashboardService {
    */
   async getKPIs(branchId?: string) {
     const cacheKey = `dashboard:kpis:${branchId || 'all'}`;
-    return this.cached(cacheKey, 60, () => this._computeKPIs(branchId)); // 60s cache
-  }
-
-  private async _computeKPIs(branchId?: string) {
-    const branchFilter: Prisma.ContractWhereInput = branchId ? { branchId } : {};
-    const productBranchFilter: Prisma.ProductWhereInput = branchId ? { branchId } : {};
-
-    const [
-      totalContracts,
-      activeContracts,
-      overdueContracts,
-      defaultContracts,
-      completedContracts,
-      totalProducts,
-      inStockProducts,
-    ] = await Promise.all([
-      this.prisma.contract.count({ where: { deletedAt: null, ...branchFilter } }),
-      this.prisma.contract.count({ where: { status: 'ACTIVE', deletedAt: null, ...branchFilter } }),
-      this.prisma.contract.count({ where: { status: 'OVERDUE', deletedAt: null, ...branchFilter } }),
-      this.prisma.contract.count({ where: { status: 'DEFAULT', deletedAt: null, ...branchFilter } }),
-      this.prisma.contract.count({ where: { status: 'COMPLETED', deletedAt: null, ...branchFilter } }),
-      this.prisma.product.count({ where: { deletedAt: null, ...productBranchFilter } }),
-      this.prisma.product.count({ where: { status: 'IN_STOCK', deletedAt: null, ...productBranchFilter } }),
-    ]);
-
-    // Financial aggregates
-    const [receivables, todayPayments] = await Promise.all([
-      this.prisma.payment.aggregate({
-        where: {
-          status: { in: ['PENDING', 'OVERDUE', 'PARTIALLY_PAID'] },
-          contract: { deletedAt: null, status: { in: ['ACTIVE', 'OVERDUE', 'DEFAULT'] }, ...branchFilter },
-        },
-        _sum: { amountDue: true, amountPaid: true, lateFee: true },
-      }),
-      this.prisma.payment.aggregate({
-        where: {
-          paidDate: (() => {
-            const now = new Date();
-            const startOfDay = new Date(now.getFullYear(), now.getMonth(), now.getDate());
-            const endOfDay = new Date(now.getFullYear(), now.getMonth(), now.getDate() + 1);
-            return { gte: startOfDay, lt: endOfDay };
-          })(),
-          status: 'PAID',
-          contract: { deletedAt: null, ...branchFilter },
-        },
-        _sum: { amountPaid: true },
-        _count: true,
-      }),
-    ]);
-
-    const totalReceivable = new Prisma.Decimal(receivables._sum.amountDue ?? 0)
-      .sub(new Prisma.Decimal(receivables._sum.amountPaid ?? 0)).toNumber();
-    const totalLateFees = new Prisma.Decimal(receivables._sum.lateFee ?? 0).toNumber();
-    const overdueRate = totalContracts > 0
-      ? ((overdueContracts + defaultContracts) / totalContracts * 100).toFixed(1)
-      : '0.0';
-
-    // MoM comparison: current month vs previous month
-    const now = new Date();
-    const thisMonthStart = new Date(now.getFullYear(), now.getMonth(), 1);
-    const prevMonthStart = new Date(now.getFullYear(), now.getMonth() - 1, 1);
-
-    const [
-      currentMonthContracts,
-      prevMonthContracts,
-      currentMonthOverdue,
-      prevMonthOverdue,
-      currentMonthStock,
-      prevMonthStock,
-    ] = await Promise.all([
-      this.prisma.contract.count({
-        where: { createdAt: { gte: thisMonthStart }, deletedAt: null, ...branchFilter },
-      }),
-      this.prisma.contract.count({
-        where: { createdAt: { gte: prevMonthStart, lt: thisMonthStart }, deletedAt: null, ...branchFilter },
-      }),
-      this.prisma.contract.count({
-        where: { status: { in: ['OVERDUE', 'DEFAULT'] }, createdAt: { gte: thisMonthStart }, deletedAt: null, ...branchFilter },
-      }),
-      this.prisma.contract.count({
-        where: { status: { in: ['OVERDUE', 'DEFAULT'] }, createdAt: { gte: prevMonthStart, lt: thisMonthStart }, deletedAt: null, ...branchFilter },
-      }),
-      this.prisma.product.count({
-        where: { status: 'IN_STOCK', createdAt: { gte: thisMonthStart }, deletedAt: null, ...productBranchFilter },
-      }),
-      this.prisma.product.count({
-        where: { status: 'IN_STOCK', createdAt: { gte: prevMonthStart, lt: thisMonthStart }, deletedAt: null, ...productBranchFilter },
-      }),
-    ]);
-
-    const computeMoM = (current: number, prev: number): number | null =>
-      prev > 0 ? ((current - prev) / prev) * 100 : null;
-
-    return {
-      contracts: { total: totalContracts, active: activeContracts, overdue: overdueContracts, default: defaultContracts, completed: completedContracts },
-      products: { total: totalProducts, inStock: inStockProducts },
-      financial: {
-        totalReceivable,
-        totalLateFees,
-        todayPayments: new Prisma.Decimal(todayPayments._sum.amountPaid ?? 0).toNumber(),
-        todayPaymentCount: todayPayments._count || 0,
-      },
-      overdueRate: Number(overdueRate),
-      contractsMoM: computeMoM(currentMonthContracts, prevMonthContracts),
-      overdueMoM: computeMoM(currentMonthOverdue, prevMonthOverdue),
-      stockMoM: computeMoM(currentMonthStock, prevMonthStock),
-    };
+    return this.cached(cacheKey, 60, () => this.overview.computeKPIs(branchId)); // 60s cache
   }
 
   /**
    * Monthly trend: new contracts vs payments received (last 12 months)
-   * OPTIMIZED: 2 batch queries instead of 24 sequential queries
    */
   async getMonthlyTrend(branchId?: string) {
-    const now = new Date();
-    const startDate = new Date(now.getFullYear(), now.getMonth() - 11, 1);
-    const branchFilter = branchId ? { branchId } : {};
-
-    // 2 batch queries instead of 24 sequential ones
-    const [contracts, payments] = await Promise.all([
-      this.prisma.contract.findMany({
-        where: { createdAt: { gte: startDate }, deletedAt: null, ...branchFilter },
-        select: { createdAt: true },
-      }),
-      this.prisma.payment.findMany({
-        where: {
-          paidDate: { gte: startDate },
-          status: 'PAID',
-          contract: { deletedAt: null, ...branchFilter },
-        },
-        select: { paidDate: true, amountPaid: true },
-      }),
-    ]);
-
-    // Group results by month in JS (fast, in-memory)
-    const months: { month: string; newContracts: number; paymentsReceived: number }[] = [];
-    for (let i = 11; i >= 0; i--) {
-      const start = new Date(now.getFullYear(), now.getMonth() - i, 1);
-      const end = new Date(now.getFullYear(), now.getMonth() - i + 1, 1);
-      const monthLabel = start.toLocaleDateString('th-TH', { year: '2-digit', month: 'short' });
-
-      const newContracts = contracts.filter(
-        (c) => c.createdAt >= start && c.createdAt < end,
-      ).length;
-
-      const paymentsReceived = payments
-        .filter((p) => p.paidDate && p.paidDate >= start && p.paidDate < end)
-        .reduce((sum, p) => sum.add(new Prisma.Decimal(p.amountPaid ?? 0)), new Prisma.Decimal(0))
-        .toNumber();
-
-      months.push({ month: monthLabel, newContracts, paymentsReceived });
-    }
-
-    return months;
+    return this.overview.getMonthlyTrend(branchId);
   }
 
   /**
    * Top 10 overdue customers
    */
   async getTopOverdue(branchId?: string) {
-    const branchFilter = branchId ? { branchId } : {};
-
-    const contracts = await this.prisma.contract.findMany({
-      where: { status: { in: ['OVERDUE', 'DEFAULT'] }, deletedAt: null, ...branchFilter },
-      include: {
-        customer: { select: { id: true, name: true, phone: true } },
-        branch: { select: { name: true } },
-        payments: {
-          where: { status: { in: ['PENDING', 'OVERDUE', 'PARTIALLY_PAID'] } },
-          select: { amountDue: true, amountPaid: true, lateFee: true, dueDate: true },
-        },
-      },
-      take: 20,
-    });
-
-    const results = contracts.map((c) => {
-      const totalOutstanding = c.payments.reduce(
-        (sum, p) => sum
-          .add(new Prisma.Decimal(p.amountDue ?? 0))
-          .sub(new Prisma.Decimal(p.amountPaid ?? 0))
-          .add(new Prisma.Decimal(p.lateFee ?? 0)),
-        new Prisma.Decimal(0),
-      ).toNumber();
-      const oldestDue = c.payments.length > 0
-        ? c.payments.reduce((oldest, p) => {
-            const d = new Date(p.dueDate);
-            return d < oldest ? d : oldest;
-          }, new Date(c.payments[0].dueDate))
-        : new Date();
-      const daysOverdue = calculateDaysOverdue(oldestDue);
-
-      return {
-        contractNumber: c.contractNumber,
-        customer: c.customer,
-        branch: c.branch.name,
-        status: c.status,
-        overdueInstallments: c.payments.length,
-        totalOutstanding,
-        daysOverdue,
-      };
-    });
-
-    return results
-      .sort((a, b) => b.totalOutstanding - a.totalOutstanding)
-      .slice(0, 10);
+    return this.collections.getTopOverdue(branchId);
   }
 
   /**
    * Contract status distribution
-   * OPTIMIZED: 1 groupBy query instead of 7 separate count queries
    */
   async getStatusDistribution(branchId?: string) {
-    const branchFilter = branchId ? { branchId } : {};
-
-    const groups = await this.prisma.contract.groupBy({
-      by: ['status'],
-      where: { deletedAt: null, ...branchFilter },
-      _count: true,
-    });
-
-    return groups.map((g) => ({
-      status: g.status,
-      count: g._count,
-    }));
+    return this.overview.getStatusDistribution(branchId);
   }
 
   /**
    * Branch comparison summary
-   * OPTIMIZED: batch queries instead of N+1 per branch
    */
   async getBranchComparison() {
-    const branches = await this.prisma.branch.findMany({
-      where: { isActive: true, deletedAt: null },
-      select: {
-        id: true,
-        name: true,
-        _count: { select: { contracts: true, products: true, users: true } },
-      },
-    });
-
-    const branchIds = branches.map((b) => b.id);
-    const monthStart = new Date(new Date().getFullYear(), new Date().getMonth(), 1);
-
-    // 2 batch queries for all branches instead of 2 per branch
-    const [overdueByBranch] = await Promise.all([
-      this.prisma.contract.groupBy({
-        by: ['branchId'],
-        where: {
-          branchId: { in: branchIds },
-          status: { in: ['OVERDUE', 'DEFAULT'] },
-          deletedAt: null,
-        },
-        _count: true,
-      }),
-      this.prisma.payment.groupBy({
-        by: ['contractId'],
-        where: {
-          paidDate: { gte: monthStart },
-          status: 'PAID',
-          contract: { branchId: { in: branchIds }, deletedAt: null },
-        },
-        _sum: { amountPaid: true },
-      }),
-    ]);
-
-    // Build payment totals per branch via a second lightweight query
-    const paidPayments = await this.prisma.payment.findMany({
-      where: {
-        paidDate: { gte: monthStart },
-        status: 'PAID',
-        contract: { branchId: { in: branchIds }, deletedAt: null },
-      },
-      select: { amountPaid: true, contract: { select: { branchId: true } } },
-    });
-
-    const paymentTotalByBranch = new Map<string, number>();
-    for (const p of paidPayments) {
-      const bid = p.contract.branchId;
-      paymentTotalByBranch.set(
-        bid,
-        new Prisma.Decimal(paymentTotalByBranch.get(bid) ?? 0)
-          .add(new Prisma.Decimal(p.amountPaid ?? 0))
-          .toNumber(),
-      );
-    }
-
-    const overdueMap = new Map(overdueByBranch.map((o) => [o.branchId, o._count]));
-
-    return branches.map((branch) => ({
-      id: branch.id,
-      name: branch.name,
-      contracts: branch._count.contracts,
-      products: branch._count.products,
-      users: branch._count.users,
-      overdueContracts: overdueMap.get(branch.id) || 0,
-      monthlyPayments: paymentTotalByBranch.get(branch.id) || 0,
-    }));
+    return this.overview.getBranchComparison();
   }
 
   /**
    * Monthly revenue summary for current month
    */
   async getMonthlyRevenue(branchId?: string) {
-    const now = new Date();
-    const monthStart = new Date(now.getFullYear(), now.getMonth(), 1);
-    const monthEnd = new Date(now.getFullYear(), now.getMonth() + 1, 1);
-    const branchFilter = branchId ? { branchId } : {};
-
-    const [paidPayments, lateFeeAgg] = await Promise.all([
-      this.prisma.payment.findMany({
-        where: {
-          paidDate: { gte: monthStart, lt: monthEnd },
-          status: 'PAID',
-          contract: { deletedAt: null, ...branchFilter },
-        },
-        select: {
-          amountPaid: true,
-          contract: { select: { interestTotal: true, totalMonths: true } },
-        },
-      }),
-      this.prisma.payment.aggregate({
-        where: {
-          paidDate: { gte: monthStart, lt: monthEnd },
-          contract: { deletedAt: null, ...branchFilter },
-        },
-        _sum: { lateFee: true },
-      }),
-    ]);
-
-    const totalPayments = paidPayments.reduce(
-      (sum, p) => sum.add(new Prisma.Decimal(p.amountPaid ?? 0)),
-      new Prisma.Decimal(0),
-    ).toNumber();
-    const interestIncome = paidPayments.reduce((sum, p) => {
-      const monthlyInterest = new Prisma.Decimal(p.contract.interestTotal ?? 0)
-        .div(p.contract.totalMonths);
-      return sum.add(monthlyInterest);
-    }, new Prisma.Decimal(0));
-
-    return {
-      totalPayments,
-      interestIncome: Math.round(interestIncome.toNumber()),
-      lateFeeIncome: new Prisma.Decimal(lateFeeAgg._sum.lateFee ?? 0).toNumber(),
-      paymentCount: paidPayments.length,
-    };
+    return this.overview.getMonthlyRevenue(branchId);
   }
 
   /**
    * Aging summary: overdue payments grouped by age buckets
    */
   async getAgingSummary(branchId?: string) {
-    const now = new Date();
-    const branchFilter = branchId
-      ? { contract: { branchId, deletedAt: null } }
-      : { contract: { deletedAt: null } };
-
-    const overduePayments = await this.prisma.payment.findMany({
-      where: {
-        status: { in: ['PENDING', 'OVERDUE', 'PARTIALLY_PAID'] },
-        dueDate: { lt: now },
-        ...branchFilter,
-      },
-      select: { amountDue: true, amountPaid: true, lateFee: true, dueDate: true },
-    });
-
-    const bucketDefs = [
-      { range: '1-30', min: 1, max: 30, color: 'green' },
-      { range: '31-60', min: 31, max: 60, color: 'yellow' },
-      { range: '61-90', min: 61, max: 90, color: 'orange' },
-      { range: '90+', min: 91, max: Infinity, color: 'red' },
-    ];
-
-    const buckets = bucketDefs.map((def) => {
-      const items = overduePayments.filter((p) => {
-        const days = calculateDaysElapsed(p.dueDate, now);
-        return days >= def.min && days <= def.max;
-      });
-      const amount = items.reduce(
-        (s, p) => s.add(new Prisma.Decimal(p.amountDue ?? 0)).sub(new Prisma.Decimal(p.amountPaid ?? 0)),
-        new Prisma.Decimal(0),
-      ).toNumber();
-      return { range: def.range, count: items.length, amount, color: def.color };
-    });
-
-    const totalCount = buckets.reduce((s, b) => s + b.count, 0);
-    const totalAmount = buckets.reduce((s, b) => s + b.amount, 0);
-
-    return { buckets, total: { count: totalCount, amount: totalAmount } };
+    return this.collections.getAgingSummary(branchId);
   }
 
   /**
    * SLA metrics: contract approval time + pending approvals > 20 min
    */
   async getSlaMetrics(branchId?: string) {
-    const branchFilter = branchId ? { branchId } : {};
-    const threshold20min = new Date(Date.now() - 20 * 60 * 1000);
-
-    // Fetch contracts reviewed in last 30 days for approval time stats
-    const thirtyDaysAgo = new Date(Date.now() - 30 * 24 * 60 * 60 * 1000);
-    const [reviewedContracts, pendingContracts] = await Promise.all([
-      this.prisma.contract.findMany({
-        where: {
-          deletedAt: null,
-          reviewedAt: { not: null, gte: thirtyDaysAgo },
-          ...branchFilter,
-        },
-        select: { createdAt: true, reviewedAt: true },
-      }),
-      this.prisma.contract.findMany({
-        where: {
-          deletedAt: null,
-          workflowStatus: { in: ['PENDING_REVIEW', 'CREATING'] },
-          reviewedAt: null,
-          ...branchFilter,
-        },
-        select: {
-          contractNumber: true,
-          createdAt: true,
-          customer: { select: { name: true } },
-        },
-      }),
-    ]);
-
-    // Calculate approval times in minutes
-    const approvalMinutes = reviewedContracts
-      .map((c) => (c.reviewedAt!.getTime() - c.createdAt.getTime()) / (1000 * 60))
-      .sort((a, b) => a - b);
-
-    const sampleCount = approvalMinutes.length;
-    const avgMinutes = sampleCount > 0
-      ? approvalMinutes.reduce((s, m) => s + m, 0) / sampleCount
-      : 0;
-    const p50Minutes = sampleCount > 0
-      ? approvalMinutes[Math.floor(sampleCount * 0.5)]
-      : 0;
-    const p90Minutes = sampleCount > 0
-      ? approvalMinutes[Math.floor(sampleCount * 0.9)]
-      : 0;
-
-    // Pending contracts with their wait times
-    const pendingList = pendingContracts.map((c) => ({
-      contractNumber: c.contractNumber,
-      customerName: c.customer?.name || '-',
-      pendingMinutes: Math.round((Date.now() - c.createdAt.getTime()) / (1000 * 60)),
-    }));
-
-    const over20MinList = pendingList.filter((c) => c.pendingMinutes > 20);
-
-    return {
-      approvalTime: {
-        avgMinutes: Math.round(avgMinutes * 10) / 10,
-        p50Minutes: Math.round(p50Minutes * 10) / 10,
-        p90Minutes: Math.round(p90Minutes * 10) / 10,
-        sampleCount,
-      },
-      pendingApprovals: {
-        totalCount: pendingContracts.length,
-        over20MinCount: over20MinList.length,
-        contracts: over20MinList,
-      },
-    };
+    return this.ops.getSlaMetrics(branchId);
   }
 
   /**
@@ -522,259 +113,14 @@ export class DashboardService {
    */
   async getAlerts(branchId?: string) {
     const cacheKey = `dashboard:alerts:${branchId || 'all'}`;
-    return this.cached(cacheKey, 30, () => this._computeAlerts(branchId)); // 30s cache
-  }
-
-  private async _computeAlerts(branchId?: string) {
-    const branchFilter = branchId ? { branchId } : {};
-
-    const [
-      overdueCount,
-      defaultCount,
-      pendingContractsCount,
-      pendingEvidenceCount,
-      activeStockAlerts,
-    ] = await Promise.all([
-      this.prisma.contract.count({
-        where: { status: 'OVERDUE', deletedAt: null, ...branchFilter },
-      }),
-      this.prisma.contract.count({
-        where: { status: 'DEFAULT', deletedAt: null, ...branchFilter },
-      }),
-      this.prisma.contract.count({
-        where: {
-          deletedAt: null,
-          workflowStatus: { in: ['PENDING_REVIEW', 'CREATING'] },
-          reviewedAt: null,
-          ...branchFilter,
-        },
-      }),
-      this.prisma.paymentEvidence.count({
-        where: {
-          status: 'PENDING_REVIEW',
-          contract: { deletedAt: null, ...branchFilter },
-        },
-      }),
-      this.prisma.stockAlert.count({
-        where: {
-          status: 'ACTIVE',
-          ...(branchId ? { reorderPoint: { branchId } } : {}),
-        },
-      }),
-    ]);
-
-    const alerts: {
-      type: string;
-      severity: 'critical' | 'warning' | 'info';
-      message: string;
-      link: string;
-      count: number;
-    }[] = [];
-
-    const totalOverdue = overdueCount + defaultCount;
-    if (totalOverdue > 0) {
-      alerts.push({
-        type: 'overdue',
-        severity: totalOverdue >= 10 ? 'critical' : 'warning',
-        message: `มี ${totalOverdue} สัญญาค้างชำระ`,
-        link: '/overdue',
-        count: totalOverdue,
-      });
-    }
-
-    if (activeStockAlerts > 0) {
-      alerts.push({
-        type: 'low_stock',
-        severity: 'warning',
-        message: `สินค้าใกล้หมด ${activeStockAlerts} รายการ`,
-        link: '/stock?tab=alerts',
-        count: activeStockAlerts,
-      });
-    }
-
-    if (pendingContractsCount > 0) {
-      alerts.push({
-        type: 'pending_contracts',
-        severity: 'info',
-        message: `มี ${pendingContractsCount} สัญญารออนุมัติ`,
-        link: '/contracts',
-        count: pendingContractsCount,
-      });
-    }
-
-    if (pendingEvidenceCount > 0) {
-      alerts.push({
-        type: 'payment_mismatch',
-        severity: pendingEvidenceCount >= 5 ? 'warning' : 'info',
-        message: `มี ${pendingEvidenceCount} สลิปรอตรวจสอบ`,
-        link: '/slip-review',
-        count: pendingEvidenceCount,
-      });
-    }
-
-    return alerts;
+    return this.cached(cacheKey, 30, () => this.ops.computeAlerts(branchId)); // 30s cache
   }
 
   /**
    * Staff performance: sales metrics (current month) + recent activity (last 7 days)
    */
   async getStaffPerformance(branchId?: string) {
-    const now = new Date();
-    const monthStart = new Date(now.getFullYear(), now.getMonth(), 1);
-    const monthEnd = new Date(now.getFullYear(), now.getMonth() + 1, 1);
-    const weekAgo = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);
-    const branchFilter = branchId ? { branchId } : {};
-
-    // Sales metrics: contracts this month grouped by salesperson via Postgres groupBy
-    // (replaces former findMany + JS reduce — old path loaded full monthly contract
-    // set with deep includes; new path returns ~N rows where N = salesperson count).
-    const baseWhere = {
-      createdAt: { gte: monthStart, lt: monthEnd },
-      deletedAt: null,
-      ...branchFilter,
-    } as const;
-
-    const [aggregates, overdueAgg] = await Promise.all([
-      this.prisma.contract.groupBy({
-        by: ['salespersonId', 'branchId'],
-        where: baseWhere,
-        _count: { _all: true },
-        _sum: { sellingPrice: true },
-      }),
-      this.prisma.contract.groupBy({
-        by: ['salespersonId'],
-        where: { ...baseWhere, status: { in: ['OVERDUE', 'DEFAULT'] } },
-        _count: { _all: true },
-      }),
-    ]);
-
-    // Enrich names + branch names in two batched queries (vs N+1 in old loop).
-    const salespersonIds = Array.from(
-      new Set(aggregates.map((a) => a.salespersonId).filter((id): id is string => !!id)),
-    );
-    const branchIds = Array.from(
-      new Set(aggregates.map((a) => a.branchId).filter((id): id is string => !!id)),
-    );
-
-    const [users, branches] = await Promise.all([
-      salespersonIds.length
-        ? this.prisma.user.findMany({
-            where: { id: { in: salespersonIds } },
-            select: { id: true, name: true },
-          })
-        : Promise.resolve([] as { id: string; name: string }[]),
-      branchIds.length
-        ? this.prisma.branch.findMany({
-            where: { id: { in: branchIds } },
-            select: { id: true, name: true },
-          })
-        : Promise.resolve([] as { id: string; name: string }[]),
-    ]);
-
-    const userNameMap = new Map(users.map((u) => [u.id, u.name]));
-    const branchNameMap = new Map(branches.map((b) => [b.id, b.name]));
-    const overdueMap = new Map(
-      overdueAgg
-        .filter((o) => !!o.salespersonId)
-        .map((o) => [o.salespersonId as string, o._count._all]),
-    );
-
-    // Collapse [salespersonId, branchId] rows → one row per salesperson.
-    // Matches old semantics: first contract's branch wins for the salesperson.
-    type Bucket = {
-      name: string;
-      branch: string;
-      totalContracts: number;
-      totalSales: number;
-      overdueCount: number;
-    };
-    const staffMap = new Map<string, Bucket>();
-    for (const a of aggregates) {
-      if (!a.salespersonId) continue;
-      const key = a.salespersonId;
-      const existing = staffMap.get(key);
-      const sellingSum = new Prisma.Decimal(a._sum.sellingPrice ?? 0).toNumber();
-      if (existing) {
-        existing.totalContracts += a._count._all;
-        existing.totalSales = new Prisma.Decimal(existing.totalSales)
-          .add(new Prisma.Decimal(a._sum.sellingPrice ?? 0))
-          .toNumber();
-      } else {
-        staffMap.set(key, {
-          name: userNameMap.get(key) ?? '-',
-          branch: branchNameMap.get(a.branchId) ?? '-',
-          totalContracts: a._count._all,
-          totalSales: sellingSum,
-          overdueCount: overdueMap.get(key) ?? 0,
-        });
-      }
-    }
-
-    const salesMetrics = Array.from(staffMap.entries())
-      .map(([id, data]) => ({
-        salespersonId: id,
-        ...data,
-        overdueRate:
-          data.totalContracts > 0
-            ? Number(((data.overdueCount / data.totalContracts) * 100).toFixed(1))
-            : 0,
-      }))
-      .sort((a, b) => b.totalSales - a.totalSales);
-
-    // Recent activity: contracts created + payments recorded in last 7 days
-    const [recentContracts, recentPayments] = await Promise.all([
-      this.prisma.contract.findMany({
-        where: { createdAt: { gte: weekAgo }, deletedAt: null, ...branchFilter },
-        select: {
-          id: true,
-          contractNumber: true,
-          sellingPrice: true,
-          createdAt: true,
-          salesperson: { select: { name: true } },
-          customer: { select: { name: true } },
-        },
-        orderBy: { createdAt: 'desc' },
-        take: 10,
-      }),
-      this.prisma.payment.findMany({
-        where: {
-          paidDate: { gte: weekAgo },
-          status: 'PAID',
-          contract: { deletedAt: null, ...branchFilter },
-        },
-        select: {
-          id: true,
-          amountPaid: true,
-          paidDate: true,
-          recordedBy: { select: { name: true } },
-          contract: { select: { contractNumber: true, customer: { select: { name: true } } } },
-        },
-        orderBy: { paidDate: 'desc' },
-        take: 10,
-      }),
-    ]);
-
-    const recentActivity = [
-      ...recentContracts.map((c) => ({
-        id: c.id,
-        type: 'contract_created' as const,
-        userName: c.salesperson.name,
-        description: `สร้างสัญญา ${c.contractNumber} — ${c.customer.name}`,
-        amount: new Prisma.Decimal(c.sellingPrice ?? 0).toNumber(),
-        createdAt: c.createdAt.toISOString(),
-      })),
-      ...recentPayments.map((p) => ({
-        id: p.id,
-        type: 'payment_recorded' as const,
-        userName: p.recordedBy?.name || '-',
-        description: `บันทึกชำระ ${p.contract.contractNumber} — ${p.contract.customer.name}`,
-        amount: new Prisma.Decimal(p.amountPaid ?? 0).toNumber(),
-        createdAt: p.paidDate?.toISOString() || new Date().toISOString(),
-      })),
-    ].sort((a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime())
-      .slice(0, 20);
-
-    return { salesMetrics, recentActivity };
+    return this.ops.getStaffPerformance(branchId);
   }
 
   /**
@@ -787,288 +133,11 @@ export class DashboardService {
    */
   async getCollectionMetrics(branchId?: string) {
     const cacheKey = `dashboard:collection-metrics:${branchId || 'all'}`;
-    return this.cached(cacheKey, 60, () => this._computeCollectionMetrics(branchId)); // 60s cache
-  }
-
-  private async _computeCollectionMetrics(branchId?: string) {
-    const now = new Date();
-    const branchContractFilter = branchId
-      ? { contract: { branchId, deletedAt: null } }
-      : { contract: { deletedAt: null } };
-
-    // ── Aging buckets ────────────────────────────────────────────────────────
-    const overduePayments = await this.prisma.payment.findMany({
-      where: {
-        status: { in: ['PENDING', 'OVERDUE', 'PARTIALLY_PAID'] },
-        dueDate: { lt: now },
-        ...branchContractFilter,
-      },
-      select: { amountDue: true, amountPaid: true, dueDate: true },
-    });
-
-    const agingBucketDefs = [
-      { label: '1-7 วัน', min: 1, max: 7 },
-      { label: '8-14 วัน', min: 8, max: 14 },
-      { label: '15-30 วัน', min: 15, max: 30 },
-      { label: '31-60 วัน', min: 31, max: 60 },
-      { label: '61-90 วัน', min: 61, max: 90 },
-      { label: '90+ วัน', min: 91, max: Infinity },
-    ];
-
-    const agingBuckets = agingBucketDefs.map((def) => {
-      const items = overduePayments.filter((p) => {
-        const days = calculateDaysElapsed(p.dueDate, now);
-        return days >= def.min && days <= def.max;
-      });
-      const amount = items
-        .reduce(
-          (s, p) =>
-            s
-              .add(new Prisma.Decimal(p.amountDue ?? 0))
-              .sub(new Prisma.Decimal(p.amountPaid ?? 0)),
-          new Prisma.Decimal(0),
-        )
-        .toNumber();
-      return { label: def.label, min: def.min, max: def.max === Infinity ? 999999 : def.max, count: items.length, amount };
-    });
-
-    // ── Collection rate ──────────────────────────────────────────────────────
-    const thisMonthStart = new Date(now.getFullYear(), now.getMonth(), 1);
-    const thisMonthEnd = new Date(now.getFullYear(), now.getMonth() + 1, 1);
-    const prevMonthStart = new Date(now.getFullYear(), now.getMonth() - 1, 1);
-
-    const [
-      thisMonthDue,
-      thisMonthCollected,
-      prevMonthDue,
-      prevMonthCollected,
-    ] = await Promise.all([
-      this.prisma.payment.aggregate({
-        where: {
-          dueDate: { gte: thisMonthStart, lt: thisMonthEnd },
-          ...branchContractFilter,
-        },
-        _sum: { amountDue: true },
-      }),
-      this.prisma.payment.aggregate({
-        where: {
-          status: 'PAID',
-          paidDate: { gte: thisMonthStart, lt: thisMonthEnd },
-          ...branchContractFilter,
-        },
-        _sum: { amountPaid: true },
-        _count: true,
-      }),
-      this.prisma.payment.aggregate({
-        where: {
-          dueDate: { gte: prevMonthStart, lt: thisMonthStart },
-          ...branchContractFilter,
-        },
-        _sum: { amountDue: true },
-      }),
-      this.prisma.payment.aggregate({
-        where: {
-          status: 'PAID',
-          paidDate: { gte: prevMonthStart, lt: thisMonthStart },
-          ...branchContractFilter,
-        },
-        _sum: { amountPaid: true },
-      }),
-    ]);
-
-    const currentDue = new Prisma.Decimal(thisMonthDue._sum.amountDue ?? 0);
-    const currentCollectedAmount = new Prisma.Decimal(thisMonthCollected._sum.amountPaid ?? 0);
-    const prevDue = new Prisma.Decimal(prevMonthDue._sum.amountDue ?? 0);
-    const prevCollectedAmount = new Prisma.Decimal(prevMonthCollected._sum.amountPaid ?? 0);
-
-    const currentRate = currentDue.gt(0)
-      ? currentCollectedAmount.div(currentDue).mul(100).toDecimalPlaces(1).toNumber()
-      : 0;
-    const lastMonthRate = prevDue.gt(0)
-      ? prevCollectedAmount.div(prevDue).mul(100).toDecimalPlaces(1).toNumber()
-      : 0;
-    const mom = Number((currentRate - lastMonthRate).toFixed(1));
-
-    // ── Top delinquent customers ─────────────────────────────────────────────
-    const branchSql = branchId ? Prisma.sql`AND c.branch_id = ${branchId}` : Prisma.empty;
-
-    const topDelinquent = await this.prisma.$queryRaw<
-      {
-        customerId: string;
-        customerName: string;
-        totalOverdue: number;
-        contractCount: number;
-      }[]
-    >(Prisma.sql`
-      SELECT
-        cu.id AS "customerId",
-        cu.name AS "customerName",
-        SUM(p.amount_due - p.amount_paid) AS "totalOverdue",
-        COUNT(DISTINCT c.id) AS "contractCount"
-      FROM customers cu
-      JOIN contracts c ON c.customer_id = cu.id AND c.deleted_at IS NULL
-      JOIN payments p ON p.contract_id = c.id
-        AND p.status IN ('PENDING', 'OVERDUE', 'PARTIALLY_PAID')
-        AND p.due_date < NOW()
-      WHERE cu.deleted_at IS NULL
-        ${branchSql}
-      GROUP BY cu.id, cu.name
-      ORDER BY SUM(p.amount_due - p.amount_paid) DESC
-      LIMIT 10
-    `);
-
-    const topDelinquentMapped = topDelinquent.map((r) => ({
-      customerId: r.customerId,
-      customerName: r.customerName,
-      totalOverdue: new Prisma.Decimal(r.totalOverdue ?? 0).toNumber(),
-      contractCount: Number(r.contractCount),
-    }));
-
-    // ── Channel effectiveness ────────────────────────────────────────────────
-    const thirtyDaysAgo = new Date(now.getTime() - 30 * 24 * 60 * 60 * 1000);
-
-    const channelRows = await this.prisma.$queryRaw<
-      {
-        channel: string;
-        totalSent: number;
-        ledToPayment: number;
-      }[]
-    >(Prisma.sql`
-      SELECT
-        da.channel,
-        COUNT(da.id) AS "totalSent",
-        COUNT(p.id) FILTER (
-          WHERE p.status = 'PAID'
-            AND p.paid_date >= da.executed_at
-            AND p.paid_date <= da.executed_at + INTERVAL '7 days'
-        ) AS "ledToPayment"
-      FROM dunning_actions da
-      JOIN contracts c ON c.id = da.contract_id AND c.deleted_at IS NULL
-      LEFT JOIN payments p ON p.contract_id = da.contract_id
-      WHERE da.status IN ('SENT', 'DELIVERED')
-        AND da.executed_at >= ${thirtyDaysAgo}
-        AND da.deleted_at IS NULL
-        ${branchSql}
-      GROUP BY da.channel
-      ORDER BY da.channel
-    `);
-
-    const channelEffectiveness = channelRows.map((r) => ({
-      channel: r.channel,
-      totalSent: Number(r.totalSent),
-      ledToPayment: Number(r.ledToPayment),
-    }));
-
-    return {
-      agingBuckets,
-      collectionRate: {
-        current: currentRate,
-        lastMonth: lastMonthRate,
-        mom,
-      },
-      collected: {
-        thisMonth: currentCollectedAmount.toNumber(),
-        count: thisMonthCollected._count || 0,
-      },
-      topDelinquent: topDelinquentMapped,
-      channelEffectiveness,
-    };
+    return this.cached(cacheKey, 60, () => this.collections.computeCollectionMetrics(branchId)); // 60s cache
   }
 
   async getWatchList(branchId?: string) {
     const cacheKey = `dashboard:watch-list:${branchId || 'all'}`;
-    return this.cached(cacheKey, 120, () => this._computeWatchList(branchId)); // 2min cache
-  }
-
-  private async _computeWatchList(branchId?: string) {
-    const branchFilter = branchId ? Prisma.sql`AND c.branch_id = ${branchId}` : Prisma.empty;
-
-    const rows = await this.prisma.$queryRaw<
-      {
-        customerId: string;
-        customerName: string;
-        customerPhone: string;
-        contractId: string;
-        contractNumber: string;
-        latePaymentCount: number;
-        partialPaymentCount: number;
-        hadDunningReset: boolean;
-        nextDueDate: Date | null;
-        nextAmountDue: number | null;
-      }[]
-    >(Prisma.sql`
-      SELECT
-        cu.id AS "customerId",
-        cu.name AS "customerName",
-        cu.phone AS "customerPhone",
-        c.id AS "contractId",
-        c.contract_number AS "contractNumber",
-        COUNT(p.id) FILTER (
-          WHERE p.paid_date IS NOT NULL AND p.paid_date::date > p.due_date::date
-        ) AS "latePaymentCount",
-        COUNT(p.id) FILTER (WHERE p.status = 'PARTIALLY_PAID') AS "partialPaymentCount",
-        (c.dunning_last_action_at IS NOT NULL AND c.dunning_stage = 'NONE') AS "hadDunningReset",
-        MIN(p2.due_date) AS "nextDueDate",
-        MIN(p2.amount_due) AS "nextAmountDue"
-      FROM customers cu
-      JOIN contracts c ON c.customer_id = cu.id AND c.deleted_at IS NULL
-      LEFT JOIN payments p ON p.contract_id = c.id
-      LEFT JOIN payments p2 ON p2.contract_id = c.id AND p2.status IN ('PENDING', 'OVERDUE')
-      WHERE cu.deleted_at IS NULL
-        AND c.status = 'ACTIVE'
-        ${branchFilter}
-      GROUP BY cu.id, c.id
-      HAVING
-        COUNT(p.id) FILTER (
-          WHERE p.paid_date IS NOT NULL AND p.paid_date::date > p.due_date::date
-        ) >= 2
-        OR COUNT(p.id) FILTER (WHERE p.status = 'PARTIALLY_PAID') >= 1
-        OR (c.dunning_last_action_at IS NOT NULL AND c.dunning_stage = 'NONE')
-      ORDER BY
-        (
-          LEAST(COUNT(p.id) FILTER (
-            WHERE p.paid_date IS NOT NULL AND p.paid_date::date > p.due_date::date
-          ), 5)
-          + COUNT(p.id) FILTER (WHERE p.status = 'PARTIALLY_PAID') * 2
-          + CASE WHEN c.dunning_last_action_at IS NOT NULL AND c.dunning_stage = 'NONE' THEN 3 ELSE 0 END
-        ) DESC
-      LIMIT 20
-    `);
-
-    const watchList = rows.map((r) => {
-      const late = Number(r.latePaymentCount);
-      const partial = Number(r.partialPaymentCount);
-      const dunningReset = Boolean(r.hadDunningReset);
-      const score = Math.min(late, 5) + partial * 2 + (dunningReset ? 3 : 0);
-      const riskLevel: 'HIGH' | 'MEDIUM' | 'LOW' = score >= 5 ? 'HIGH' : score >= 3 ? 'MEDIUM' : 'LOW';
-
-      const reasons: string[] = [];
-      if (late >= 2) reasons.push(`ชำระล่าช้า ${late} ครั้ง`);
-      if (partial >= 1) reasons.push(`จ่ายไม่ครบ ${partial} ครั้ง`);
-      if (dunningReset) reasons.push('เคยถูกติดตามหนี้แล้ว reset');
-
-      return {
-        customerId: r.customerId,
-        customerName: r.customerName,
-        customerPhone: r.customerPhone,
-        contractId: r.contractId,
-        contractNumber: r.contractNumber,
-        riskScore: score,
-        riskLevel,
-        reasons,
-        nextDueDate: r.nextDueDate,
-        nextAmountDue: r.nextAmountDue ? new Prisma.Decimal(r.nextAmountDue).toNumber() : null,
-      };
-    });
-
-    const highCount = watchList.filter((w) => w.riskLevel === 'HIGH').length;
-    const mediumCount = watchList.filter((w) => w.riskLevel === 'MEDIUM').length;
-
-    return {
-      total: watchList.length,
-      highCount,
-      mediumCount,
-      watchList,
-    };
+    return this.cached(cacheKey, 120, () => this.collections.computeWatchList(branchId)); // 2min cache
   }
 }
