+++ apps/api/src/modules/dashboard/services/dashboard-overview.service.ts
@@ -1,0 +1,301 @@
+import { Injectable } from '@nestjs/common';
+import { PrismaService } from '../../../prisma/prisma.service';
+import { Prisma } from '@prisma/client';
+
+@Injectable()
+export class DashboardOverviewService {
+  constructor(private prisma: PrismaService) {}
+
+  async computeKPIs(branchId?: string) {
+    const branchFilter: Prisma.ContractWhereInput = branchId ? { branchId } : {};
+    const productBranchFilter: Prisma.ProductWhereInput = branchId ? { branchId } : {};
+
+    const [
+      totalContracts,
+      activeContracts,
+      overdueContracts,
+      defaultContracts,
+      completedContracts,
+      totalProducts,
+      inStockProducts,
+    ] = await Promise.all([
+      this.prisma.contract.count({ where: { deletedAt: null, ...branchFilter } }),
+      this.prisma.contract.count({ where: { status: 'ACTIVE', deletedAt: null, ...branchFilter } }),
+      this.prisma.contract.count({ where: { status: 'OVERDUE', deletedAt: null, ...branchFilter } }),
+      this.prisma.contract.count({ where: { status: 'DEFAULT', deletedAt: null, ...branchFilter } }),
+      this.prisma.contract.count({ where: { status: 'COMPLETED', deletedAt: null, ...branchFilter } }),
+      this.prisma.product.count({ where: { deletedAt: null, ...productBranchFilter } }),
+      this.prisma.product.count({ where: { status: 'IN_STOCK', deletedAt: null, ...productBranchFilter } }),
+    ]);
+
+    // Financial aggregates
+    const [receivables, todayPayments] = await Promise.all([
+      this.prisma.payment.aggregate({
+        where: {
+          status: { in: ['PENDING', 'OVERDUE', 'PARTIALLY_PAID'] },
+          contract: { deletedAt: null, status: { in: ['ACTIVE', 'OVERDUE', 'DEFAULT'] }, ...branchFilter },
+        },
+        _sum: { amountDue: true, amountPaid: true, lateFee: true },
+      }),
+      this.prisma.payment.aggregate({
+        where: {
+          paidDate: (() => {
+            const now = new Date();
+            const startOfDay = new Date(now.getFullYear(), now.getMonth(), now.getDate());
+            const endOfDay = new Date(now.getFullYear(), now.getMonth(), now.getDate() + 1);
+            return { gte: startOfDay, lt: endOfDay };
+          })(),
+          status: 'PAID',
+          contract: { deletedAt: null, ...branchFilter },
+        },
+        _sum: { amountPaid: true },
+        _count: true,
+      }),
+    ]);
+
+    const totalReceivable = new Prisma.Decimal(receivables._sum.amountDue ?? 0)
+      .sub(new Prisma.Decimal(receivables._sum.amountPaid ?? 0)).toNumber();
+    const totalLateFees = new Prisma.Decimal(receivables._sum.lateFee ?? 0).toNumber();
+    const overdueRate = totalContracts > 0
+      ? ((overdueContracts + defaultContracts) / totalContracts * 100).toFixed(1)
+      : '0.0';
+
+    // MoM comparison: current month vs previous month
+    const now = new Date();
+    const thisMonthStart = new Date(now.getFullYear(), now.getMonth(), 1);
+    const prevMonthStart = new Date(now.getFullYear(), now.getMonth() - 1, 1);
+
+    const [
+      currentMonthContracts,
+      prevMonthContracts,
+      currentMonthOverdue,
+      prevMonthOverdue,
+      currentMonthStock,
+      prevMonthStock,
+    ] = await Promise.all([
+      this.prisma.contract.count({
+        where: { createdAt: { gte: thisMonthStart }, deletedAt: null, ...branchFilter },
+      }),
+      this.prisma.contract.count({
+        where: { createdAt: { gte: prevMonthStart, lt: thisMonthStart }, deletedAt: null, ...branchFilter },
+      }),
+      this.prisma.contract.count({
+        where: { status: { in: ['OVERDUE', 'DEFAULT'] }, createdAt: { gte: thisMonthStart }, deletedAt: null, ...branchFilter },
+      }),
+      this.prisma.contract.count({
+        where: { status: { in: ['OVERDUE', 'DEFAULT'] }, createdAt: { gte: prevMonthStart, lt: thisMonthStart }, deletedAt: null, ...branchFilter },
+      }),
+      this.prisma.product.count({
+        where: { status: 'IN_STOCK', createdAt: { gte: thisMonthStart }, deletedAt: null, ...productBranchFilter },
+      }),
+      this.prisma.product.count({
+        where: { status: 'IN_STOCK', createdAt: { gte: prevMonthStart, lt: thisMonthStart }, deletedAt: null, ...productBranchFilter },
+      }),
+    ]);
+
+    const computeMoM = (current: number, prev: number): number | null =>
+      prev > 0 ? ((current - prev) / prev) * 100 : null;
+
+    return {
+      contracts: { total: totalContracts, active: activeContracts, overdue: overdueContracts, default: defaultContracts, completed: completedContracts },
+      products: { total: totalProducts, inStock: inStockProducts },
+      financial: {
+        totalReceivable,
+        totalLateFees,
+        todayPayments: new Prisma.Decimal(todayPayments._sum.amountPaid ?? 0).toNumber(),
+        todayPaymentCount: todayPayments._count || 0,
+      },
+      overdueRate: Number(overdueRate),
+      contractsMoM: computeMoM(currentMonthContracts, prevMonthContracts),
+      overdueMoM: computeMoM(currentMonthOverdue, prevMonthOverdue),
+      stockMoM: computeMoM(currentMonthStock, prevMonthStock),
+    };
+  }
+
+  /**
+   * Monthly trend: new contracts vs payments received (last 12 months)
+   * OPTIMIZED: 2 batch queries instead of 24 sequential queries
+   */
+  async getMonthlyTrend(branchId?: string) {
+    const now = new Date();
+    const startDate = new Date(now.getFullYear(), now.getMonth() - 11, 1);
+    const branchFilter = branchId ? { branchId } : {};
+
+    // 2 batch queries instead of 24 sequential ones
+    const [contracts, payments] = await Promise.all([
+      this.prisma.contract.findMany({
+        where: { createdAt: { gte: startDate }, deletedAt: null, ...branchFilter },
+        select: { createdAt: true },
+      }),
+      this.prisma.payment.findMany({
+        where: {
+          paidDate: { gte: startDate },
+          status: 'PAID',
+          contract: { deletedAt: null, ...branchFilter },
+        },
+        select: { paidDate: true, amountPaid: true },
+      }),
+    ]);
+
+    // Group results by month in JS (fast, in-memory)
+    const months: { month: string; newContracts: number; paymentsReceived: number }[] = [];
+    for (let i = 11; i >= 0; i--) {
+      const start = new Date(now.getFullYear(), now.getMonth() - i, 1);
+      const end = new Date(now.getFullYear(), now.getMonth() - i + 1, 1);
+      const monthLabel = start.toLocaleDateString('th-TH', { year: '2-digit', month: 'short' });
+
+      const newContracts = contracts.filter(
+        (c) => c.createdAt >= start && c.createdAt < end,
+      ).length;
+
+      const paymentsReceived = payments
+        .filter((p) => p.paidDate && p.paidDate >= start && p.paidDate < end)
+        .reduce((sum, p) => sum.add(new Prisma.Decimal(p.amountPaid ?? 0)), new Prisma.Decimal(0))
+        .toNumber();
+
+      months.push({ month: monthLabel, newContracts, paymentsReceived });
+    }
+
+    return months;
+  }
+
+  /**
+   * Contract status distribution
+   * OPTIMIZED: 1 groupBy query instead of 7 separate count queries
+   */
+  async getStatusDistribution(branchId?: string) {
+    const branchFilter = branchId ? { branchId } : {};
+
+    const groups = await this.prisma.contract.groupBy({
+      by: ['status'],
+      where: { deletedAt: null, ...branchFilter },
+      _count: true,
+    });
+
+    return groups.map((g) => ({
+      status: g.status,
+      count: g._count,
+    }));
+  }
+
+  /**
+   * Branch comparison summary
+   * OPTIMIZED: batch queries instead of N+1 per branch
+   */
+  async getBranchComparison() {
+    const branches = await this.prisma.branch.findMany({
+      where: { isActive: true, deletedAt: null },
+      select: {
+        id: true,
+        name: true,
+        _count: { select: { contracts: true, products: true, users: true } },
+      },
+    });
+
+    const branchIds = branches.map((b) => b.id);
+    const monthStart = new Date(new Date().getFullYear(), new Date().getMonth(), 1);
+
+    // 2 batch queries for all branches instead of 2 per branch
+    const [overdueByBranch] = await Promise.all([
+      this.prisma.contract.groupBy({
+        by: ['branchId'],
+        where: {
+          branchId: { in: branchIds },
+          status: { in: ['OVERDUE', 'DEFAULT'] },
+          deletedAt: null,
+        },
+        _count: true,
+      }),
+      this.prisma.payment.groupBy({
+        by: ['contractId'],
+        where: {
+          paidDate: { gte: monthStart },
+          status: 'PAID',
+          contract: { branchId: { in: branchIds }, deletedAt: null },
+        },
+        _sum: { amountPaid: true },
+      }),
+    ]);
+
+    // Build payment totals per branch via a second lightweight query
+    const paidPayments = await this.prisma.payment.findMany({
+      where: {
+        paidDate: { gte: monthStart },
+        status: 'PAID',
+        contract: { branchId: { in: branchIds }, deletedAt: null },
+      },
+      select: { amountPaid: true, contract: { select: { branchId: true } } },
+    });
+
+    const paymentTotalByBranch = new Map<string, number>();
+    for (const p of paidPayments) {
+      const bid = p.contract.branchId;
+      paymentTotalByBranch.set(
+        bid,
+        new Prisma.Decimal(paymentTotalByBranch.get(bid) ?? 0)
+          .add(new Prisma.Decimal(p.amountPaid ?? 0))
+          .toNumber(),
+      );
+    }
+
+    const overdueMap = new Map(overdueByBranch.map((o) => [o.branchId, o._count]));
+
+    return branches.map((branch) => ({
+      id: branch.id,
+      name: branch.name,
+      contracts: branch._count.contracts,
+      products: branch._count.products,
+      users: branch._count.users,
+      overdueContracts: overdueMap.get(branch.id) || 0,
+      monthlyPayments: paymentTotalByBranch.get(branch.id) || 0,
+    }));
+  }
+
+  /**
+   * Monthly revenue summary for current month
+   */
+  async getMonthlyRevenue(branchId?: string) {
+    const now = new Date();
+    const monthStart = new Date(now.getFullYear(), now.getMonth(), 1);
+    const monthEnd = new Date(now.getFullYear(), now.getMonth() + 1, 1);
+    const branchFilter = branchId ? { branchId } : {};
+
+    const [paidPayments, lateFeeAgg] = await Promise.all([
+      this.prisma.payment.findMany({
+        where: {
+          paidDate: { gte: monthStart, lt: monthEnd },
+          status: 'PAID',
+          contract: { deletedAt: null, ...branchFilter },
+        },
+        select: {
+          amountPaid: true,
+          contract: { select: { interestTotal: true, totalMonths: true } },
+        },
+      }),
+      this.prisma.payment.aggregate({
+        where: {
+          paidDate: { gte: monthStart, lt: monthEnd },
+          contract: { deletedAt: null, ...branchFilter },
+        },
+        _sum: { lateFee: true },
+      }),
+    ]);
+
+    const totalPayments = paidPayments.reduce(
+      (sum, p) => sum.add(new Prisma.Decimal(p.amountPaid ?? 0)),
+      new Prisma.Decimal(0),
+    ).toNumber();
+    const interestIncome = paidPayments.reduce((sum, p) => {
+      const monthlyInterest = new Prisma.Decimal(p.contract.interestTotal ?? 0)
+        .div(p.contract.totalMonths);
+      return sum.add(monthlyInterest);
+    }, new Prisma.Decimal(0));
+
+    return {
+      totalPayments,
+      interestIncome: Math.round(interestIncome.toNumber()),
+      lateFeeIncome: new Prisma.Decimal(lateFeeAgg._sum.lateFee ?? 0).toNumber(),
+      paymentCount: paidPayments.length,
+    };
+  }
+}
